feat(api): Phase 5.4 - migrate ml_ready to canonical _ml_ready module
