+++ python/zedda/__init__.py
@@ -1155,277 +1155,10 @@ def warnings(
 
 
 # ─────────────────────────────────────────────────────────────────
-#  _ml_readiness_score() — dedicated ML readiness scoring
-#
-#  More ML-specific than _quality_score() — penalizes issues that
-#  directly impact model training quality:
-#    - Null columns (moderate + severe)
-#    - ID-like columns (wasted features)
-#    - High-cardinality strings (need encoding)
-#    - Extreme outliers
-#    - Multicollinearity (correlated pairs)
-# ─────────────────────────────────────────────────────────────────
-def _ml_readiness_score(p: Any) -> int:
-    """Compute a 0-100 ML readiness score from the profile object."""
-    score = 100
-
-    for col in p.columns:
-        # Penalize columns with notable nulls
-        if col.null_pct > 50:
-            score -= 15  # Too sparse — can't trust imputation
-        elif col.null_pct > 5:
-            score -= 10  # Needs imputation but recoverable
-
-        # Penalize ID-like columns (useless features)
-        if col.type_str == "int" and col.unique_pct > 95:
-            score -= 5
-
-        # Penalize high-cardinality strings (ID-like)
-        if (
-            col.type_str in ("str", "unknown")
-            and p.num_rows > 0
-            and col.unique_approx > p.num_rows * 0.8
-        ):
-            score -= 5
-        # Penalize moderate-cardinality strings (needs encoding)
-        elif col.type_str in ("str", "unknown") and col.unique_approx > 100:
-            score -= 3
-
-        # Penalize extreme outliers
-        if (
-            col.type_str in ("int", "float")
-            and col.mean > 0
-            and col.unique_approx > 5
-            and col.val_max > 10
-            and col.val_max > col.mean * 10
-            and "ratio" not in col.name.lower()
-            and "pct" not in col.name.lower()
-        ):
-            score -= 5
-
-    # Penalize strongly correlated pairs (multicollinearity)
-    for cr in p.correlations:
-        if abs(cr.r) >= 0.95:
-            score -= 10
-            break  # Count once
-
-    return max(0, min(100, score))
-
-
-def _section_header(title: str, width: int = 55) -> str:
-    """Return a ──── Title ──── style Rich-formatted section header."""
-    ch = _safe_symbol("─", "-")
-    left = ch * 14
-    right_len = max(1, width - 14 - len(title) - 2)
-    right = ch * right_len
-    return f"[dim]{left}[/dim] [bold]{title}[/bold] [dim]{right}[/dim]"
-
-
-# ─────────────────────────────────────────────────────────────────
 #  ml_ready() — ML readiness check with premium terminal UI
-#
-#  Shows:
-#    1. Version header + mode label
-#    2. Scan timing
-#    3. Visual score bar (█░ style)
-#    4. Grouped issues with inline fix code
-#    5. "Looks Good" section for clean features
-#    6. Copy-paste fix code block
-#    7. Summary footer
 # ─────────────────────────────────────────────────────────────────
-def ml_ready(
-    path,
-    target: str | None = None,
-    sample_size: int | None = None,
-    correlate: bool = False,
-) -> None:
-    """
-    Check if a dataset is ready for Machine Learning.
+from ._ml_ready import ml_ready
 
-    Provides a score (0-100) with a visual progress bar, Target Column section,
-    and Feature Verdict Table with plain-English action recommendations.
-
-    Args:
-        path (str): Path to a ``.csv``, ``.parquet``, or ``.arrow`` file.
-        target (str, optional): Target/label column name.
-        sample_size (int, optional): Max rows to sample.
-
-    Example::
-
-        import zedda as zd
-        zd.ml_ready("data.csv", target="churn")
-    """
-    resolved_path, is_in_memory = _resolve_input(path)
-    try:
-        t0 = time.perf_counter()
-        p = _scan_wrapper(resolved_path, sample_size=sample_size, correlate=correlate)
-        total_ms = (time.perf_counter() - t0) * 1000
-
-        if not _RICH_AVAILABLE or _console is None:
-            raise ZeddaError(
-                "Rich is required for terminal output. Install with: pip install rich"
-            )
-
-        file_name = (
-            "<DataFrame>"
-            if is_in_memory
-            else (Path(path).name if isinstance(path, (str, Path)) else "<DataFrame>")
-        )
-        scan_str = (
-            f"{total_ms / 1000:.1f} sec" if total_ms >= 10_000 else f"{total_ms:.0f} ms"
-        )
-
-        crit_sym = _safe_symbol("✗", "[X]")
-        warn_sym = _safe_symbol("⚠", "[!]")
-        check_sym = _safe_symbol("✓", "[OK]")
-        arrow_r = _safe_symbol("→", "->")
-        arrow_lr = _safe_symbol("↔", "<->")
-        bullet = _safe_symbol("·", "-")
-
-        # ── Header ─────────────────────────────────────────────────
-        _console.print(
-            f"\n[bold blue]zedda[/bold blue] [dim]v{__version__}[/dim]  {bullet}  "
-            f"[bold]ml_ready mode[/bold]\n"
-        )
-        _console.print(
-            f"[dim]Scanning[/dim]   [cyan]{file_name}[/cyan]   ...  {scan_str}\n"
-        )
-
-        # ── ML Readiness Score ─────────────────────────────────────
-        base_score = _quality_score(p)
-        score = _ml_readiness_score(p)
-        bar = _render_quality_bar(score)
-        color, label = _quality_label(score)
-
-        _console.print(_section_header("ML Readiness Score"))
-        _console.print(f"  Base Data Quality score : {base_score} / 100")
-        score_diff = score - base_score
-        diff_str = f"({score_diff:+} pts from base)" if score_diff != 0 else ""
-        _console.print(
-            f"  [{color}]ML Readiness score      : {score} / 100  {bar}  {label}  {diff_str}[/{color}]\n"
-        )
-
-        # ── Target Column ──────────────────────────────────────────
-        _console.print(_section_header("Target Column"))
-        detected_target = None
-        if target:
-            target_col = next((c for c in p.columns if c.name == target), None)
-            if target_col:
-                detected_target = target_col.name
-                _console.print(
-                    f"  Target       : [bold cyan]'{rich_escape(target)}'[/bold cyan] (using specified target)"
-                )
-            else:
-                _console.print(
-                    f"  Target       : [bold yellow]'{rich_escape(target)}'[/bold yellow] (specified target not found)"
-                )
-        else:
-            # Auto-detect binary target candidate
-            binary_cand = next(
-                (
-                    c
-                    for c in p.columns
-                    if c.type_str in ("int", "float")
-                    and c.val_min == 0
-                    and c.val_max == 1
-                    and c.unique_approx <= 2
-                ),
-                None,
-            )
-            if binary_cand:
-                detected_target = binary_cand.name
-                _console.print(
-                    f"  Target       : [bold cyan]'{rich_escape(binary_cand.name)}'[/bold cyan] (auto-detected binary classification)"
-                )
-            else:
-                _console.print(
-                    "  Target       : [dim]None specified (unsupervised / feature audit mode)[/dim]"
-                )
-        _console.print()
-
-        # ── Feature Verdict Table ──────────────────────────────────
-        _console.print(_section_header("Feature Verdict Table"))
-        table_verdict = Table(
-            show_header=True,
-            header_style="bold white on blue",
-            border_style="dim",
-            box=box.SIMPLE_HEAVY,
-            padding=(0, 1),
-        )
-        table_verdict.add_column("Feature", style="bold cyan", min_width=12)
-        table_verdict.add_column("Verdict", min_width=14)
-        table_verdict.add_column("Reason", min_width=25)
-        table_verdict.add_column("Action", min_width=20)
-
-        drop_cols = []
-        keep_cols = []
-
-        for col in p.columns:
-            if detected_target and col.name == detected_target:
-                table_verdict.add_row(
-                    rich_escape(col.name),
-                    "[bold green]TARGET[/bold green]",
-                    "Target / label column",
-                    "Keep as target",
-                )
-                continue
-
-            issues_found = _detect_column_issues(col, p)
-            if not issues_found:
-                table_verdict.add_row(
-                    rich_escape(col.name),
-                    "[bold green]KEEP as-is[/bold green]",
-                    "Clean distribution",
-                    "Keep as-is",
-                )
-                keep_cols.append(col.name)
-            else:
-                issue = issues_found[0]
-                action = _get_fix_action(col, issue)
-                if action["action_type"] == "drop":
-                    table_verdict.add_row(
-                        rich_escape(col.name),
-                        "[bold red]DROP[/bold red]",
-                        action["message"],
-                        "Drop column",
-                    )
-                    drop_cols.append(col.name)
-                else:
-                    act_desc = (
-                        "Impute median"
-                        if "impute" in action["action_type"]
-                        or "missing" in issue.get("message", "").lower()
-                        else (
-                            "One-hot encode"
-                            if "encode" in action["action_type"]
-                            or "string" in issue.get("message", "").lower()
-                            or "cardinality" in issue.get("message", "").lower()
-                            else "Fix issue"
-                        )
-                    )
-                    table_verdict.add_row(
-                        rich_escape(col.name),
-                        "[bold yellow]KEEP after fix[/bold yellow]",
-                        action["message"],
-                        act_desc,
-                    )
-                    keep_cols.append(col.name)
-
-        _console.print(table_verdict)
-        _console.print()
-
-        # ── Footer ─────────────────────────────────────────────────
-        recommended = len(keep_cols)
-        _console.print(
-            f"  [dim]Recommended features for training: {recommended} of {p.num_cols} candidate columns.[/dim]"
-        )
-        _console.print(
-            f'  [dim]Run zd.fix("{file_name}") to generate executable pipeline code.[/dim]\n'
-        )
-
-    finally:
-        if is_in_memory:
-            _cleanup_temp(resolved_path)
 
 
 # ─────────────────────────────────────────────────────────────────
