+++ tests/python/test_ml_ready.py
@@ -24,3 +24,34 @@ def test_ml_ready_type_coercion():
     )
     # Should not crash on type coercion (Task 2 fix)
     zd.ml_ready(df)
+
+
+def test_ml_ready_canonical_module():
+    from zedda._ml_ready import ml_ready as _ml_ready_fn
+    assert zd.ml_ready is _ml_ready_fn
+
+
+def test_ml_ready_with_target_and_output(tmp_path):
+    import io
+    from contextlib import redirect_stdout
+
+    df = pd.DataFrame(
+        {
+            "id": range(50),
+            "feature1": [1.0 if i % 2 == 0 else 2.0 for i in range(50)],
+            "target": [0 if i % 2 == 0 else 1 for i in range(50)],
+        }
+    )
+    p = tmp_path / "train.csv"
+    df.to_csv(p, index=False)
+
+    f = io.StringIO()
+    with redirect_stdout(f):
+        zd.ml_ready(str(p), target="target")
+    out = f.getvalue()
+
+    assert "ML Readiness Score" in out
+    assert "Target Column" in out
+    assert "Feature Verdict Table" in out
+    assert "TARGET" in out
+
